chore: 게임관련 함수 소켓에서 분리 중
