+++ src/game/objects/player.ts
@@ -28,26 +28,25 @@ export class Character {
         this.id = id
         this.position = position
         this.velocity = [0, 0, 0]
-        this.bodyColor = Character.generateRandomHexColor()
-        this.hairColor = Character.generateRandomHexColor()
-        this.bellyColor = Character.generateRandomHexColor()
+        this.bodyColor = this.generateRandomHexColor()
+        this.hairColor = this.generateRandomHexColor()
+        this.bellyColor = this.generateRandomHexColor()
         this.isOnGround = true
         this.directions = { up: false, down: false, left: false, right: false }
         this.hasTail = Math.random() > 0.5
         this.shift = false
         this.isBeingStolen = false
 
-        // Cannon.js body 설정
         const body = new CANNON.Body({
             mass: 1,
             position: new CANNON.Vec3(position[0], position[1], position[2]),
             material: material,
+            shape: new CANNON.Sphere(1),
         })
-        body.addShape(new CANNON.Sphere(1)) // 캐릭터의 크기 설정
         this.cannonBody = body
     }
 
-    static generateRandomHexColor(): string {
+    private generateRandomHexColor(): string {
         const color = Math.floor(Math.random() * 16777215).toString(16)
         return '#' + color.padStart(6, '0')
     }
