+++ src/game/rounds/index.ts
@@ -1,0 +1,3 @@
+import { CommonRound } from './common'
+import { TailTagRound } from './tail-tag'
+export { CommonRound, TailTagRound }
